Remove name and email from expense and revenue lists
This aligns all printouts of user real names.
Some instances printed the email and some the UUID.
The detail pages now always prints the real name, and the UUID.
If the user does not have a real name, the UUID is shown as a fallback option, instead of presenting an empty cell.
All real names have been removed from the economy app, and will only be shown in the backoffice app.